% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.adonis.R
\name{fast.adonis}
\alias{fast.adonis}
\title{Calculate R2 for analysis of variance using distance matrices}
\usage{
fast.adonis(
  formula,
  data = NULL,
  permutations = 999,
  boot.times = 100,
  boot.se = "WCB",
  boot.sample.size = NULL,
  weights = NULL,
  order_list = NULL,
  by = "terms",
  parallel = getOption("mc.cores")
)
}
\arguments{
\item{formula}{Model Formula. The LHS must be a matrix that is computed from a dissimilarity matrix. For example, given a Bray Curtis dissimilarity matrix D, LHS= -0.5*D^2. The RHS defines the independent variables. These can be continuous variables or factors.}

\item{data}{the data frame for the independent variables.}

\item{permutations}{number of permutations}

\item{boot.times}{times for bootstrapping}

\item{boot.se}{the name of any method for bootstraps. Bootstraps are used to calculate standard errors for R2. 'boot.se = WCB' is within cluster bootstrap. Clusters are first created by weights.'boot.se = AS' is booting samples directly by weights. By default, 'boot.se = WCB'.}

\item{boot.sample.size}{sample size for bootstrap}

\item{weights}{If weights is NULL or a vector of 1s, \code{fast.adonis} conducts both permutations and bootstraps. Otherwise, fast.adonis` does not conduct permutations.}

\item{order_list}{a list of different orders of input variables. For example, an input formula is RHS ~ A+B+C. The order_list can be list(c(3,2,1),c(2,3,1)). The output will be three tabs. The first is the analysis table for formula, RHS ~ A+B+C. The second is the table for formula, RHS ~ C+B+A. The last is the table for formula, RHS ~ B+C+A.}

\item{by}{'by=terms' will assess significance for each term (sequentially from first to last). 'by=margin' will assess the marginal effects of the terms (each marginal term analysis in a model without all other variables). If 'by=margin' is set, "order_list" will be set to NULL.}

\item{parallel}{number of parallel processes. With parallel = 1 uses ordinary, non-parallel processing. The parallel processing is done with the parallel package.}
}
\value{
An anova table
}
\description{
\code{fast.adonis} is a R function that is used for analysis of variance using distance matrices.
It can be used for both weighted and unweighted data. For unweighted data, it performs
similarly with R functions 'adonis' and 'adonis2' from package(vegan). Besides a permutation
test, \code{fast.adonis} provides a standard error for R2. The standard error can by computed by
within cluster bootstraps or bootstrapping samples by weights. For unweighted data,
\code{fast.adonis} does not provide permutation tests.
}
